+++ nirvana/tests/test_geometry.py
@@ -118,3 +118,4 @@ def test_deriv_projected_polar():
         'Azimuth derivative does not match finite difference'
 
 
+